Move useful test functions into silx.utils.testutils
