+++ silx/gui/plot/_utils/test/test_ticklayout.py
@@ -27,13 +27,13 @@ from __future__ import absolute_import, division, unicode_literals
 
 __authors__ = ["T. Vincent"]
 __license__ = "MIT"
-__date__ = "18/10/2016"
+__date__ = "17/01/2018"
 
 
 import unittest
 import numpy
 
-from silx.test.utils import ParametricTestCase
+from silx.utils.testutils import ParametricTestCase
 
 from silx.gui.plot._utils import ticklayout
