+++ silx/gui/plot/test/testColors.py
@@ -26,13 +26,13 @@
 
 __authors__ = ["T. Vincent"]
 __license__ = "MIT"
-__date__ = "05/12/2016"
+__date__ = "17/01/2018"
 
 
 import numpy
 
 import unittest
-from silx.test.utils import ParametricTestCase
+from silx.utils.testutils import ParametricTestCase
 
 from silx.gui.plot import Colors
 from silx.gui.plot.Colormap import Colormap
