+++ silx/gui/plot/test/testComplexImageView.py
@@ -26,14 +26,14 @@
 
 __authors__ = ["T. Vincent"]
 __license__ = "MIT"
-__date__ = "14/09/2017"
+__date__ = "17/01/2018"
 
 
 import unittest
 import logging
 import numpy
 
-from silx.test.utils import ParametricTestCase
+from silx.utils.testutils import ParametricTestCase
 from silx.gui.plot import ComplexImageView
 
 from .utils import PlotWidgetTestCase
