+++ silx/image/test/test_shapes.py
@@ -27,14 +27,14 @@
 
 __authors__ = ["T. Vincent"]
 __license__ = "MIT"
-__date__ = "15/05/2017"
+__date__ = "17/01/2018"
 
 
 import logging
 import unittest
 import numpy
 
-from silx.test.utils import ParametricTestCase
+from silx.utils.testutils import ParametricTestCase
 from silx.image import shapes
 
 _logger = logging.getLogger(__name__)
